Allows the Arc Gotham Themes to Install Under gtk-3.24
When the install instructions were followed under gtk-3.24 (the default in ubuntu 18.10),
configure failed: reporting an invalid gtk version.

It looks like the version check depends on existence of a version-named directory under
common/gnome-shell and common/gtk-3.0.

Adding symlinks for 3.24 that point to the latest concrete directories appears to bypass
the block on installation.

I've completed the install instructions on 18.10 (with gtk3.24) and the theme seems to work
fine.